requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
yarl==1.9.4
